Implement partial routing key hashing: "hash-routing-key-segment".
A new option has been implemented: "hash-routing-key-segment", that
allows the user to specify which part of the routing-key to use for
computing the hash.

Until now, when using the exchange hashing the routing key, the whole
routing-key was used to compute the hash of each message. This is fine
sometimes, but does not support all these use cases where the
routing key partitioning and the partitioning of the data don't
match. Particularly, this option adds support for those use cases where
the data partition matches a particular segment or subset of segments of
the routing key.